refactor: Rename SpatialJoinNode.Type -> SpatialJoinNode.SpatialJoinType
The name Type caused an identifier conflict during codegen for the cpp protocol.
Renaming it something unique avoids that.
