+++ presto-main-base/src/main/java/com/facebook/presto/sql/planner/iterative/rule/ExtractSpatialJoins.java
@@ -457,7 +457,7 @@ public class ExtractSpatialJoins
         return Result.ofPlanNode(new SpatialJoinNode(
                 joinNode.getSourceLocation(),
                 nodeId,
-                SpatialJoinNode.Type.fromJoinNodeType(joinNode.getType()),
+                SpatialJoinNode.SpatialJoinType.fromJoinNodeType(joinNode.getType()),
                 newLeftNode,
                 newRightNode,
                 outputVariables,
